apply MutationHelper and GraphElementBuilder.
